requirements.txt
absl-py==2.1.0
alembic==1.14.1
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
cmaes==0.11.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozendict==2.4.6
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.13.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.2.0
keras==2.13.1
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.40.0
lxml==5.3.1
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.7.2
multitasking==0.0.11
nltk==3.9.1
numba==0.57.1
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
optuna==3.2.0
packaging==23.1
pandas==2.0.3
pillow==11.1.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==1.0.0
pytz==2023.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.15.2
seaborn==0.12.2
shap==0.46.0
six==1.17.0
slicer==0.0.8
soupsieve==2.6
SQLAlchemy==2.0.24
ta==0.11.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.5.0
threadpoolctl==2.2.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.3.0
vaderSentiment==3.3.2
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.2.26
